% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_ui.R
\name{mod_home_ui_combined}
\alias{mod_home_ui_combined}
\title{Home UI for Combined Data}
\usage{
mod_home_ui_combined(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Shiny UI for combined/mixed data home screen
}
\description{
Home UI for Combined Data
}
