% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_ui.R
\name{mod_home_ui_meristic}
\alias{mod_home_ui_meristic}
\title{Home UI for Meristic Data}
\usage{
mod_home_ui_meristic(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Shiny UI for meristic home screen
}
\description{
Home UI for Meristic Data
}
