% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_ui.R
\name{mod_home_ui_morphometric}
\alias{mod_home_ui_morphometric}
\title{Home UI for Morphometric Data}
\usage{
mod_home_ui_morphometric(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Shiny UI for morphometric home screen
}
\description{
Home UI for Morphometric Data
}
