% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_ui.R
\name{sidebar_modules_header_ui}
\alias{sidebar_modules_header_ui}
\title{Sidebar Modules Header}
\usage{
sidebar_modules_header_ui()
}
\value{
Shiny UI for the Modules section header
}
\description{
A styled "Modules" section header to insert in the sidebar between the
welcome block and the module navigation links.
}
