% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_ui.R
\name{sidebar_welcome_ui}
\alias{sidebar_welcome_ui}
\title{Sidebar Welcome Panel}
\usage{
sidebar_welcome_ui()
}
\value{
Shiny UI for the sidebar links block
}
\description{
Links row displayed below the Return to Home button in the left sidebar.
}
